add AuthAccount and refactor special case models check
