+++ xrpl/models/__init__.py
@@ -1,6 +1,7 @@
 """Top-level exports for the models package."""
 from xrpl.models import amounts, currencies, requests, transactions
 from xrpl.models.amounts import *  # noqa: F401, F403
+from xrpl.models.auth_account import AuthAccount
 from xrpl.models.currencies import *  # noqa: F401, F403
 from xrpl.models.exceptions import XRPLModelException
 from xrpl.models.path import Path, PathStep
@@ -13,6 +14,7 @@ __all__ = [
     "XRPLModelException",
     "amounts",
     *amounts.__all__,
+    "AuthAccount",
     "currencies",
     *currencies.__all__,
     "requests",
