+++ xrpl/models/requests/__init__.py
@@ -1,5 +1,6 @@
 """Request models."""
-from xrpl.models.path import PathStep
+from xrpl.models.auth_account import AuthAccount
+from xrpl.models.path_step import PathStep
 from xrpl.models.requests.account_channels import AccountChannels
 from xrpl.models.requests.account_currencies import AccountCurrencies
 from xrpl.models.requests.account_info import AccountInfo
@@ -54,6 +55,7 @@ __all__ = [
     "AccountOffers",
     "AccountTx",
     "AMMInfo",
+    "AuthAccount",
     "BookOffers",
     "ChannelAuthorize",
     "ChannelVerify",
